requirements.txt
pyspark>=3.4.0
delta-spark>=2.4.0
requests>=2.28.0
